+++ TableProTests/PluginTestSources/RedshiftSchemaQueries.swift
@@ -1,0 +1,1 @@
+../../Plugins/PostgreSQLDriverPlugin/RedshiftSchemaQueries.swift